Separate between public and private
Separate public and private classes and methods, and adapt
test cases to work
