+++ snapcraft/storeapi/_client.py
