+++ snapcraft/tests/unit/commands/test_list_revisions.py
@@ -101,7 +101,7 @@ class RevisionsCommandTestCase(RevisionsCommandBaseTestCase):
         self.assertThat(str(raised), Equals(
             "Snap 'snap-test' was not found in '18' series."))
 
-    @mock.patch.object(storeapi.SCAClient, 'snap_revisions')
+    @mock.patch.object(storeapi._sca_client.SCAClient, 'snap_revisions')
     @mock.patch.object(storeapi.StoreClient, 'get_account_information')
     def test_revisions_by_unknown_arch(self, mock_account_api, mock_revisions):
         mock_revisions.return_value = {}
@@ -114,7 +114,7 @@ class RevisionsCommandTestCase(RevisionsCommandBaseTestCase):
         self.assertThat(str(raised), Equals(
             "Snap 'snap-test' for 'some-arch' was not found in '16' series."))
 
-    @mock.patch.object(storeapi.SCAClient, 'snap_revisions')
+    @mock.patch.object(storeapi._sca_client.SCAClient, 'snap_revisions')
     @mock.patch.object(storeapi.StoreClient, 'get_account_information')
     def test_revisions_by_unknown_series(self,
                                          mock_account_api, mock_revisions):
