+++ snapcraft/tests/unit/commands/test_login.py
@@ -39,7 +39,8 @@ class LoginCommandTestCase(CommandBaseTestCase):
         patcher.start()
         self.addCleanup(patcher.stop)
 
-    @mock.patch.object(storeapi.SCAClient, 'get_account_information')
+    @mock.patch.object(storeapi._sca_client.SCAClient,
+                       'get_account_information')
     @mock.patch.object(storeapi.StoreClient, 'login')
     def test_successful_login(
             self, mock_login, mock_get_account_information):
@@ -58,7 +59,8 @@ class LoginCommandTestCase(CommandBaseTestCase):
             'user@example.com', mock.ANY, acls=None, packages=None,
             channels=None, save=True)
 
-    @mock.patch.object(storeapi.SCAClient, 'get_account_information')
+    @mock.patch.object(storeapi._sca_client.SCAClient,
+                       'get_account_information')
     @mock.patch.object(storeapi.StoreClient, 'login')
     def test_successful_login_with_2fa(
             self, mock_login, mock_get_account_information):
